Clean session ID strings
tmux will automatically replace certains characters with an underscore.
For example, directories containing a period will have the period
replaced. This breaks tsm since the new session will be created, but tsm
will be unable to switch to it. For this change, we assume valid session
names match the regular expression /[a-zA-Z0-9_-]+/.